Player GameObject added with Prefab engine VFX
Mesh collider being used instead of primitive capsule collider. Model's mesh was simplified through Unity provided asset for this project with a mesh of only 16 triangles. With Unity 5, Convex must be selected in order for an object to be physics enabled. This should also be used in conjunction with the Trigger option. https://unity3d.com/learn/tutorials/projects/space-shooter-tutorial/camera-and-lighting?playlist=17147 is the next step for project.